+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Changelog
 
+**0.9.13**
+
+* AngularClient - Dashboard and Project manager minor update/fix
+* AngularClient - Search is moved to toolbar
+
 **0.9.12**
 
 * AngularClient - Dashboard fix for other users views